% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-preparation.R
\name{head.zcurve_data}
\alias{head.zcurve_data}
\title{Prints first few rows of a z-curve data object}
\usage{
\method{head}{zcurve_data}(x, ...)
}
\arguments{
\item{x}{z-curve data object}

\item{...}{Additional arguments}
}
\description{
Prints first few rows of a z-curve data object
}
\seealso{
\code{\link[=zcurve_data]{zcurve_data()}}
}
