% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{OSC.z}
\alias{OSC.z}
\title{Z-scores from subset of original studies featured in OSC 2015
reproducibility project}
\format{A vector with 90 observations}
\usage{
OSC.z
}
\description{
The dataset contains z-scores from subset of original
studies featured in psychology reproducibility project
\insertCite{osc}{zcurve}. Only z-scores from studies with unambiguous
original outcomes are supplied (eliminating 7 studies with marginally
significant results). The real replication rate for those studies is
35/90 (the whole project reports 36/97).
}
\references{
\insertAllCited{}
}
\keyword{datasets}
