% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{print.estimates.zcurve}
\alias{print.estimates.zcurve}
\title{Prints estimates from z-curve object}
\usage{
\method{print.estimates}{zcurve}(x, ...)
}
\arguments{
\item{x}{Estimate of a z-curve object}

\item{...}{Additional arguments}
}
\description{
Prints estimates from z-curve object
}
\seealso{
\code{\link[=zcurve]{zcurve()}}
}
