% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{print.summary.zcurve}
\alias{print.summary.zcurve}
\title{Prints summary object for z-curve method}
\usage{
\method{print.summary}{zcurve}(x, ...)
}
\arguments{
\item{x}{Summary of a z-curve object}

\item{...}{Additional arguments}
}
\description{
Prints summary object for z-curve method
}
\seealso{
\code{\link[=zcurve]{zcurve()}}
}
