% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{print.zcurve}
\alias{print.zcurve}
\title{Prints a fitted z-curve object}
\usage{
\method{print}{zcurve}(x, ...)
}
\arguments{
\item{x}{Fitted z-curve object}

\item{...}{Additional arguments}
}
\description{
Prints a fitted z-curve object
}
\seealso{
\code{\link[=zcurve]{zcurve()}}
}
