% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-preparation.R
\name{print.zcurve_data}
\alias{print.zcurve_data}
\title{Prints a z-curve data object}
\usage{
\method{print}{zcurve_data}(x, ...)
}
\arguments{
\item{x}{z-curve data object}

\item{...}{Additional arguments}
}
\description{
Prints a z-curve data object
}
\seealso{
\code{\link[=zcurve_data]{zcurve_data()}}
}
