% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zcurve-package.R
\docType{package}
\name{zcurve-package}
\alias{zcurve-package}
\title{zcurve: An Implementation of Z-Curves}
\description{
An implementation of z-curves - a method for estimating expected discovery and replicability rates on the bases of test-statistics of published studies. The package provides functions for fitting the density, EM, and censored EM version (Bartoš & Schimmack, 2022, \doi{10.15626/MP.2021.2720}; Schimmack & Bartoš, 2023, \doi{ 10.1371/journal.pone.0290084}), as well as the original density z-curve (Brunner & Schimmack, 2020, \doi{10.15626/MP.2018.874}). Furthermore, the package provides summarizing and plotting functions for the fitted z-curve objects. See the aforementioned articles for more information about the z-curves, expected discovery and replicability rates, validation studies, and limitations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://fbartos.github.io/zcurve/}
  \item Report bugs at \url{https://github.com/FBartos/zcurve/issues}
}

}
\author{
\strong{Maintainer}: František Bartoš \email{f.bartos96@gmail.com}

Authors:
\itemize{
  \item Ulrich Schimmack \email{ulrich.schimmack@utoronto.ca}
}

}
\keyword{internal}
