% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{zcurve}
\alias{zcurve}
\title{Fit a z-curve}
\usage{
zcurve(z, method = "EM", bootstrap = 1000, control = NULL)
}
\arguments{
\item{z}{a vector of z-scores}

\item{method}{the method to be used for fitting. Possible options are
Expectation Maximization \code{"EM"} and density \code{"density"},
defaults to \code{"EM"}.}

\item{bootstrap}{the number of bootstraps for estimating CI. To skip
bootstrap specify \code{FALSE}}

\item{control}{additional options for the fitting algorithm more details in
\link[=control_EM]{control EM} or \link[=control_density]{control density}}
}
\value{
The fitted z-curve object
}
\description{
Fit a z-curve
}
\examples{
# simulate some z-statistics
z <- abs(rnorm(300,3))

# fit an EM7 z-curve
m.EM <- zcurve(z, method = "EM", bootstrap = 100)

# or z19.9 z-curve
m.D <- zcurve(z, method = "density", bootstrap = 100)

# inspect the results
summary(m.EM)
summary(m.D)

# plot the results
plot(m.EM)
plot(m.D)

# to specify more options, set the control arguments
# ei. increase the maximum number of iterations and change alpha level
ctr1 <- list(
  "max_iter" = 9999,
  "alpha"    = .10
  )
m1.EM <- zcurve(z, method = "EM", bootstrap = 100, control = ctr1)

}
\seealso{
\code{\link[=summary.zcurve]{summary.zcurve()}}, \code{\link[=plot.zcurve]{plot.zcurve()}}, \link{control_EM}, \link{control_density}
}
