% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{zcurve_options}
\alias{zcurve_options}
\alias{zcurve.options}
\alias{zcurve.get_option}
\title{Options for the zcurve package}
\usage{
zcurve.options(...)

zcurve.get_option(name)
}
\arguments{
\item{...}{named option(s) to change - for a list of available options, see
details below.}

\item{name}{the name of the option to get the current value of - for a list of
available options, see details below.}
}
\value{
The current value of all available zcurve options (after applying any
changes specified) is returned invisibly as a named list.
}
\description{
A placeholder object and functions for the zcurve package.
(adapted from the runjags R package).
}
