% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zcurve-package.R
\docType{package}
\name{zcurve-package}
\alias{zcurve-package}
\alias{_PACKAGE}
\title{zcurve: Implementations of Z-curves}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
